test: restore auth persistence integrity
Restore the SQLite-backed auth wizard assertions traced to 43ea501f38d and consolidate duplicated message-action and temp-directory fixtures.
