+++ extensions/codex/src/app-server/test-support.ts
@@ -3,12 +3,9 @@
  * transports.
  */
 import { EventEmitter } from "node:events";
-import fs from "node:fs";
-import path from "node:path";
 import { PassThrough, Writable } from "node:stream";
 import type { EmbeddedRunAttemptParamsV2 as EmbeddedRunAttemptParams } from "openclaw/plugin-sdk/agent-harness-runtime";
 import type { Model } from "openclaw/plugin-sdk/llm";
-import { resolvePreferredOpenClawTmpDir } from "openclaw/plugin-sdk/temp-path";
 import { vi } from "vitest";
 import { CodexAppServerClient } from "./client.js";
 import type { CodexAppServerClientFactory, CodexAppServerClientOptions } from "./shared-client.js";
@@ -78,24 +75,7 @@ export function createCodexTestToolTerminalObserver(): NonNullable<
   };
 }
 
-/** Creates temp directories that are removed by the supplied test cleanup hook. */
-export function useAutoCleanupTempDirTracker(registerCleanup: (cleanup: () => void) => unknown) {
-  const dirs = new Set<string>();
-  registerCleanup(() => {
-    for (const dir of dirs) {
-      fs.rmSync(dir, { recursive: true, force: true, maxRetries: 5, retryDelay: 20 });
-    }
-    dirs.clear();
-  });
-  return {
-    dirs,
-    make(prefix: string): string {
-      const dir = fs.mkdtempSync(path.join(resolvePreferredOpenClawTmpDir(), prefix));
-      dirs.add(dir);
-      return dir;
-    },
-  };
-}
+export { useAutoCleanupTempDirTracker } from "openclaw/plugin-sdk/test-env";
 
 /** Positional naked-client injection contract confined to tests. */
 export type CodexTestAppServerClientFactory = (
