+++ src/commands/onboard-auth.test.ts
@@ -1,9 +1,13 @@
 import fs from "node:fs/promises";
-import os from "node:os";
 import path from "node:path";
 // Onboard auth tests cover provider auth setup, credential persistence, and auth-profile state.
 import { createRequireRecord } from "openclaw/plugin-sdk/test-fixtures";
 import { afterEach, describe, expect, it, vi } from "vitest";
+import {
+  createAuthTestLifecycle,
+  readAuthProfilesForAgent,
+  setupAuthTestEnv,
+} from "../../test/helpers/auth-wizard.js";
 import type { OAuthCredentials } from "../llm/utils/oauth/types.js";
 import {
   applyAuthProfileConfig,
@@ -11,11 +15,6 @@ import {
   writeOAuthCredentials,
 } from "../plugins/provider-auth-helpers.js";
 import { setTestEnvValue } from "../test-utils/env.js";
-import {
-  createAuthTestLifecycle,
-  readAuthProfilesForAgent,
-  setupAuthTestEnv,
-} from "./test-wizard-helpers.js";
 
 const providerEnvVarsById = vi.hoisted(
   (): Record<string, readonly string[]> => ({
@@ -29,62 +28,10 @@ const providerEnvVarsById = vi.hoisted(
   }),
 );
 
-vi.mock("../config/paths.js", () => ({
+vi.mock("../config/paths.js", async (importOriginal) => ({
+  ...(await importOriginal<typeof import("../config/paths.js")>()),
   resolveStateDir: () => process.env.OPENCLAW_STATE_DIR ?? "/tmp/openclaw-state",
 }));
-
-vi.mock("../agents/auth-profiles/profiles.js", async () => {
-  const fsLocal = await import("node:fs");
-  const pathLocal = await import("node:path");
-  const upsert = (params: { profileId: string; credential: unknown; agentDir?: string }) => {
-    const stateDir = process.env.OPENCLAW_STATE_DIR ?? "/tmp/openclaw-state";
-    const agentDir = params.agentDir ?? pathLocal.join(stateDir, "agents", "main", "agent");
-    const file = pathLocal.join(agentDir, "auth-profiles.json");
-    fsLocal.mkdirSync(agentDir, { recursive: true });
-    const existing = (() => {
-      try {
-        return JSON.parse(fsLocal.readFileSync(file, "utf8")) as {
-          version?: number;
-          profiles?: Record<string, unknown>;
-        };
-      } catch {
-        return { version: 1, profiles: {} };
-      }
-    })();
-    fsLocal.writeFileSync(
-      file,
-      `${JSON.stringify(
-        {
-          version: existing.version ?? 1,
-          profiles: {
-            ...existing.profiles,
-            [params.profileId]: params.credential,
-          },
-        },
-        null,
-        2,
-      )}\n`,
-    );
-  };
-  return {
-    upsertAuthProfile: upsert,
-    upsertAuthProfileWithLock: async (params: {
-      profileId: string;
-      credential: unknown;
-      agentDir?: string;
-    }) => {
-      upsert(params);
-      return { version: 1, profiles: {} };
-    },
-    upsertAuthProfileWithLockOrThrow: async (params: {
-      profileId: string;
-      credential: unknown;
-      agentDir?: string;
-    }) => {
-      upsert(params);
-    },
-  };
-});
 
 vi.mock("../agents/provider-auth-aliases.js", () => ({
   resolveProviderIdForAuth: (provider: string) => {
@@ -115,16 +62,6 @@ function expectFields(value: unknown, expected: Record<string, unknown>, label =
   return record;
 }
 
-async function expectMissingFile(readPromise: Promise<unknown>) {
-  try {
-    await readPromise;
-  } catch (error) {
-    expectFields(error, { code: "ENOENT" }, "read error");
-    return;
-  }
-  throw new Error("Expected file read to fail with ENOENT");
-}
-
 describe("writeOAuthCredentials", () => {
   const lifecycle = createAuthTestLifecycle([
     "OPENCLAW_STATE_DIR",
@@ -132,16 +69,13 @@ describe("writeOAuthCredentials", () => {
     "OPENCLAW_OAUTH_DIR",
   ]);
 
-  let tempStateDir: string;
-  const authProfilePathFor = (dir: string) => path.join(dir, "auth-profiles.json");
-
   afterEach(async () => {
     await lifecycle.cleanup();
   });
 
-  it("writes auth-profiles.json under the default agent dir", async () => {
+  it("persists OAuth credentials under the default agent SQLite store", async () => {
     const env = await setupAuthTestEnv("openclaw-oauth-");
-    lifecycle.setStateDir(env.stateDir);
+    lifecycle.track(env);
     const defaultAgentDir = path.join(env.stateDir, "agents", "main", "agent");
 
     const creds = {
@@ -161,13 +95,15 @@ describe("writeOAuthCredentials", () => {
       type: "oauth",
     });
 
-    await expectMissingFile(fs.readFile(path.join(env.agentDir, "auth-profiles.json"), "utf8"));
+    await expect(readAuthProfilesForAgent(env.agentDir)).rejects.toThrow(
+      "Expected SQLite auth profile store",
+    );
   });
 
-  it("writes OAuth credentials to all sibling agent dirs when syncSiblingAgents=true", async () => {
-    tempStateDir = await fs.mkdtemp(path.join(os.tmpdir(), "openclaw-oauth-sync-"));
-    lifecycle.setStateDir(tempStateDir);
-    setTestEnvValue("OPENCLAW_STATE_DIR", tempStateDir);
+  it("persists primary and main OAuth rows while later siblings inherit", async () => {
+    const env = await setupAuthTestEnv("openclaw-oauth-sync-");
+    lifecycle.track(env);
+    const tempStateDir = env.stateDir;
 
     const mainAgentDir = path.join(tempStateDir, "agents", "main", "agent");
     const kidAgentDir = path.join(tempStateDir, "agents", "kid", "agent");
@@ -188,23 +124,26 @@ describe("writeOAuthCredentials", () => {
       syncSiblingAgents: true,
     });
 
-    for (const dir of [mainAgentDir, kidAgentDir, workerAgentDir]) {
-      const raw = await fs.readFile(authProfilePathFor(dir), "utf8");
-      const parsed = JSON.parse(raw) as {
+    for (const dir of [mainAgentDir, kidAgentDir]) {
+      const persistedStore = await readAuthProfilesForAgent<{
         profiles?: Record<string, OAuthCredentials & { type?: string }>;
-      };
-      expectFields(parsed.profiles?.["openai:default"], {
+      }>(dir);
+      expectFields(persistedStore.profiles?.["openai:default"], {
         refresh: "refresh-sync",
         access: "access-sync",
         type: "oauth",
       });
     }
+    const inheritedSiblingStore = await readAuthProfilesForAgent<{
+      profiles?: Record<string, OAuthCredentials & { type?: string }>;
+    }>(workerAgentDir);
+    expect(inheritedSiblingStore.profiles).toEqual({});
   });
 
   it("writes OAuth credentials only to target dir by default", async () => {
-    tempStateDir = await fs.mkdtemp(path.join(os.tmpdir(), "openclaw-oauth-nosync-"));
-    lifecycle.setStateDir(tempStateDir);
-    setTestEnvValue("OPENCLAW_STATE_DIR", tempStateDir);
+    const env = await setupAuthTestEnv("openclaw-oauth-nosync-");
+    lifecycle.track(env);
+    const tempStateDir = env.stateDir;
 
     const mainAgentDir = path.join(tempStateDir, "agents", "main", "agent");
     const kidAgentDir = path.join(tempStateDir, "agents", "kid", "agent");
@@ -221,22 +160,23 @@ describe("writeOAuthCredentials", () => {
 
     await writeOAuthCredentials("openai", creds, kidAgentDir);
 
-    const kidRaw = await fs.readFile(authProfilePathFor(kidAgentDir), "utf8");
-    const kidParsed = JSON.parse(kidRaw) as {
+    const kidParsed = await readAuthProfilesForAgent<{
       profiles?: Record<string, OAuthCredentials & { type?: string }>;
-    };
+    }>(kidAgentDir);
     expectFields(kidParsed.profiles?.["openai:default"], {
       access: "access-kid",
       type: "oauth",
     });
 
-    await expectMissingFile(fs.readFile(authProfilePathFor(mainAgentDir), "utf8"));
+    await expect(readAuthProfilesForAgent(mainAgentDir)).rejects.toThrow(
+      "Expected SQLite auth profile store",
+    );
   });
 
   it("syncs siblings from explicit agentDir outside OPENCLAW_STATE_DIR", async () => {
-    tempStateDir = await fs.mkdtemp(path.join(os.tmpdir(), "openclaw-oauth-external-"));
-    lifecycle.setStateDir(tempStateDir);
-    setTestEnvValue("OPENCLAW_STATE_DIR", tempStateDir);
+    const env = await setupAuthTestEnv("openclaw-oauth-external-");
+    lifecycle.track(env);
+    const tempStateDir = env.stateDir;
 
     // Create standard-layout agents tree *outside* OPENCLAW_STATE_DIR
     const externalRoot = path.join(tempStateDir, "external", "agents");
@@ -259,10 +199,9 @@ describe("writeOAuthCredentials", () => {
 
     // All siblings under the external root should have credentials
     for (const dir of [extMain, extKid, extWorker]) {
-      const raw = await fs.readFile(authProfilePathFor(dir), "utf8");
-      const parsed = JSON.parse(raw) as {
+      const parsed = await readAuthProfilesForAgent<{
         profiles?: Record<string, OAuthCredentials & { type?: string }>;
-      };
+      }>(dir);
       expectFields(parsed.profiles?.["openai:default"], {
         refresh: "refresh-ext",
         access: "access-ext",
@@ -272,7 +211,9 @@ describe("writeOAuthCredentials", () => {
 
     // Global state dir should NOT have credentials written
     const globalMain = path.join(tempStateDir, "agents", "main", "agent");
-    await expectMissingFile(fs.readFile(authProfilePathFor(globalMain), "utf8"));
+    await expect(readAuthProfilesForAgent(globalMain)).rejects.toThrow(
+      "Expected SQLite auth profile store",
+    );
   });
 });
 
@@ -304,9 +245,16 @@ describe("upsertApiKeyProfile secret refs", () => {
     return parsed.profiles?.[profileId];
   }
 
+  async function readProfileIds(agentDir: string): Promise<string[]> {
+    const parsed = await readAuthProfilesForAgent<{
+      profiles?: Record<string, AuthProfileEntry>;
+    }>(agentDir);
+    return Object.keys(parsed.profiles ?? {}).toSorted();
+  }
+
   it("handles plaintext, ref mode, and inline env-ref provider keys", async () => {
     const env = await setupAuthTestEnv("openclaw-onboard-auth-credentials-");
-    lifecycle.setStateDir(env.stateDir);
+    lifecycle.track(env);
     process.env.MOONSHOT_API_KEY = "sk-moonshot-env"; // pragma: allowlist secret
     process.env.OPENAI_API_KEY = "sk-openai-env"; // pragma: allowlist secret
 
@@ -371,7 +319,7 @@ describe("upsertApiKeyProfile secret refs", () => {
 
   it("stores provider-specific env refs and metadata in ref mode", async () => {
     const env = await setupAuthTestEnv("openclaw-onboard-auth-credentials-provider-ref-");
-    lifecycle.setStateDir(env.stateDir);
+    lifecycle.track(env);
     process.env.CLOUDFLARE_AI_GATEWAY_API_KEY = "cf-secret"; // pragma: allowlist secret
     process.env.VOLCANO_ENGINE_API_KEY = "volcengine-secret"; // pragma: allowlist secret
     process.env.BYTEPLUS_API_KEY = "byteplus-secret"; // pragma: allowlist secret
@@ -401,6 +349,13 @@ describe("upsertApiKeyProfile secret refs", () => {
       });
     }
 
+    expect(await readProfileIds(env.agentDir)).toEqual([
+      "byteplus:default",
+      "cloudflare-ai-gateway:default",
+      "opencode-go:default",
+      "opencode:default",
+      "volcengine:default",
+    ]);
     expectFields(await readProfile(env.agentDir, "cloudflare-ai-gateway:default"), {
       keyRef: { source: "env", provider: "default", id: "CLOUDFLARE_AI_GATEWAY_API_KEY" },
       metadata: { accountId: "account-1", gatewayId: "gateway-1" },
@@ -430,7 +385,7 @@ describe("upsertApiKeyProfile", () => {
 
   it("writes to the default agent dir", async () => {
     const env = await setupAuthTestEnv("openclaw-minimax-", { agentSubdir: "custom-agent" });
-    lifecycle.setStateDir(env.stateDir);
+    lifecycle.track(env);
     const defaultAgentDir = path.join(env.stateDir, "agents", "main", "agent");
 
     upsertApiKeyProfile({ provider: "minimax", input: "sk-minimax-test" });
@@ -444,7 +399,9 @@ describe("upsertApiKeyProfile", () => {
       key: "sk-minimax-test",
     });
 
-    await expectMissingFile(fs.readFile(path.join(env.agentDir, "auth-profiles.json"), "utf8"));
+    await expect(readAuthProfilesForAgent(env.agentDir)).rejects.toThrow(
+      "Expected SQLite auth profile store",
+    );
   });
 });
