--- src/commands/test-wizard-helpers.ts
@@ -1,10 +1,0 @@
-// Re-exported auth wizard test helpers for command tests.
-// This keeps command test imports stable while the shared helper lives under test/helpers.
-
-export {
-  createAuthTestLifecycle,
-  createExitThrowingRuntime,
-  createWizardPrompter,
-  readAuthProfilesForAgent,
-  setupAuthTestEnv,
-} from "../../test/helpers/auth-wizard.js";
