+++ src/infra/outbound/message-action-routing.context.test.ts
@@ -1,136 +1,23 @@
 // Covers message-action cross-context policy, markers, and presentation
 // decoration behavior.
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { jsonResult } from "../../agents/tools/common.js";
-import type {
-  ChannelMessageActionContext,
-  ChannelPlugin,
-} from "../../channels/plugins/types.public.js";
+import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import type { OpenClawConfig } from "../../config/config.js";
-import { setActivePluginRegistry } from "../../plugins/runtime.js";
-import {
-  createChannelTestPluginBase,
-  createTestRegistry,
-} from "../../test-utils/channel-plugins.js";
 import { MessageActionDeniedError } from "./message-action-denial.js";
 import { runMessageAction } from "./message-action-runner.js";
 import {
+  createMessageActionContextFixture,
   directChatConfig,
-  directChatTestPlugin,
-  directOutbound,
-  forumTestPlugin,
   runDryAction,
   runDrySend,
   workspaceConfig,
-  workspaceTestPlugin,
 } from "./message-action-runner.test-support.js";
 
-const handleWorkspaceAction = vi.fn(async (_ctx: ChannelMessageActionContext) =>
-  jsonResult({ ok: true }),
-);
-
-const readWorkspaceTestPlugin: ChannelPlugin = {
-  ...workspaceTestPlugin,
-  actions: {
-    describeMessageTool: () => ({ actions: ["read"] }),
-    handleAction: handleWorkspaceAction,
-  },
-};
-
-const localChatTestPlugin: ChannelPlugin = {
-  ...createChannelTestPluginBase({
-    id: "localchat",
-    label: "Local Chat",
-    docsPath: "/channels/localchat",
-    capabilities: { chatTypes: ["direct", "group"], media: true },
-  }),
-  meta: {
-    id: "localchat",
-    label: "Local Chat",
-    selectionLabel: "Local Chat (local)",
-    docsPath: "/channels/localchat",
-    blurb: "Local chat test stub.",
-    aliases: ["local"],
-  },
-  outbound: directOutbound,
-  messaging: {
-    normalizeTarget: (raw) => raw.trim() || undefined,
-    targetResolver: {
-      looksLikeId: (raw) => raw.trim().length > 0,
-      hint: "<handle|chat_id:ID>",
-    },
-  },
-};
-
-const resolvedDmTestPlugin: ChannelPlugin = {
-  ...createChannelTestPluginBase({
-    id: "slackdm",
-    label: "Resolved DM",
-    capabilities: { chatTypes: ["direct"], media: true },
-  }),
-  outbound: directOutbound,
-  messaging: {
-    normalizeTarget: (raw) => {
-      const trimmed = raw.trim();
-      if (!trimmed) {
-        return undefined;
-      }
-      const userId = trimmed.replace(/^user:/i, "");
-      return /^user:/i.test(trimmed)
-        ? `user:${userId.toLowerCase()}`
-        : `channel:${trimmed.toLowerCase()}`;
-    },
-    targetResolver: {
-      looksLikeId: (raw) => /^(?:user:)?[UW][A-Z0-9]+$/i.test(raw.trim()),
-      hint: "<user:ID>",
-      resolveTarget: async ({ input }) => {
-        const userId = input.trim().replace(/^user:/i, "");
-        return /^[UW][A-Z0-9]+$/i.test(userId)
-          ? { to: userId, kind: "user", source: "normalized" }
-          : null;
-      },
-    },
-  },
-  threading: {
-    matchesToolContextTarget: ({ target, toolContext }) =>
-      target.toLowerCase() ===
-      toolContext.currentMessagingTarget?.replace(/^user:/i, "").toLowerCase(),
-  },
-};
+const contextFixture = createMessageActionContextFixture();
+const { handleWorkspaceAction } = contextFixture;
 
 describe("runMessageAction context isolation", () => {
-  beforeEach(() => {
-    setActivePluginRegistry(
-      createTestRegistry([
-        {
-          pluginId: "workspace",
-          source: "test",
-          plugin: readWorkspaceTestPlugin,
-        },
-        {
-          pluginId: "directchat",
-          source: "test",
-          plugin: directChatTestPlugin,
-        },
-        {
-          pluginId: "forum",
-          source: "test",
-          plugin: forumTestPlugin,
-        },
-        {
-          pluginId: "localchat",
-          source: "test",
-          plugin: localChatTestPlugin,
-        },
-        {
-          pluginId: "slackdm",
-          source: "test",
-          plugin: resolvedDmTestPlugin,
-        },
-      ]),
-    );
-    handleWorkspaceAction.mockClear();
-  });
+  beforeEach(() => contextFixture.setup());
+  afterEach(() => contextFixture.cleanup());
   it("uses the current conversation for an implicit read", async () => {
     await runMessageAction({
       cfg: workspaceConfig,
@@ -155,10 +42,6 @@ describe("runMessageAction context isolation", () => {
         to: "C12345678",
       },
     });
-  });
-
-  afterEach(() => {
-    setActivePluginRegistry(createTestRegistry([]));
   });
 
   it.each([
