+++ src/infra/outbound/message-action-runner.test-helpers.ts
@@ -1,6 +1,7 @@
 // Shared runner-facade test harness. These mocks isolate message-action routing,
 // execution, and send coordination from real channel and gateway runtimes.
 import { vi } from "vitest";
+import { jsonResult } from "../../agents/tools/common.js";
 import { dispatchChannelMessageAction } from "../../channels/plugins/message-action-dispatch.js";
 import type {
   ChannelMessageActionContext,
@@ -134,6 +135,70 @@ export function createAlwaysConfiguredPluginConfig(
     resolveAccount: () => account,
     isConfigured: () => true,
   };
+}
+
+export function createActionHubPluginFixture() {
+  const handleAction = vi.fn(async ({ params }: { params: Record<string, unknown> }) =>
+    jsonResult({ ok: true, params }),
+  );
+  const plugin: ChannelPlugin = {
+    id: "actionhub",
+    meta: {
+      id: "actionhub",
+      label: "Action Hub",
+      selectionLabel: "Action Hub",
+      docsPath: "/channels/actionhub",
+      blurb: "Action Hub action dispatch test plugin.",
+    },
+    capabilities: { chatTypes: ["direct", "channel"] },
+    config: createAlwaysConfiguredPluginConfig(),
+    messaging: {
+      targetPrefixes: ["actionhub", "actionhub-alias"],
+      normalizeTarget: (raw) => raw.replace(/^actionhub-alias:/i, "actionhub:"),
+      targetResolver: {
+        looksLikeId: () => true,
+      },
+    },
+    actions: {
+      describeMessageTool: () => ({
+        actions: [
+          "pin",
+          "unpin",
+          "list-pins",
+          "member-info",
+          "channel-info",
+          "edit",
+          "thread-create",
+          "thread-reply",
+        ],
+      }),
+      messageActionTargetAliases: {
+        edit: {
+          aliases: ["messageId", "chatId", "chat_id", "channel_id"],
+          deliveryTargetAliases: ["chatId", "chat_id", "channel_id"],
+        },
+        pin: {
+          aliases: ["messageId", "chatId", "chat_id", "channel_id"],
+          deliveryTargetAliases: ["chatId", "chat_id", "channel_id"],
+        },
+        unpin: {
+          aliases: ["messageId", "chatId", "chat_id", "channel_id"],
+          deliveryTargetAliases: ["chatId", "chat_id", "channel_id"],
+        },
+      },
+      supportsAction: ({ action }) =>
+        action === "pin" ||
+        action === "unpin" ||
+        action === "list-pins" ||
+        action === "member-info" ||
+        action === "channel-info" ||
+        action === "edit" ||
+        action === "thread-create" ||
+        action === "thread-reply",
+      handleAction,
+    },
+  };
+  return { handleAction, plugin };
 }
 
 export function createGatewayActionPlugin(params: {
