+++ test/helpers/auth-wizard.ts
@@ -2,9 +2,11 @@
 import fs from "node:fs/promises";
 import path from "node:path";
 import { vi } from "vitest";
+import { loadPersistedAuthProfileStore } from "../../src/agents/auth-profiles/persisted.js";
+import { clearRuntimeAuthProfileStoreSnapshots } from "../../src/agents/auth-profiles/runtime-snapshots.js";
 import type { RuntimeEnv } from "../../src/runtime.js";
-import { makeTempWorkspace } from "../../src/test-helpers/workspace.js";
 import { captureEnv } from "../../src/test-utils/env.js";
+import { createOpenClawTestState } from "../../src/test-utils/openclaw-test-state.js";
 import type { WizardPrompter } from "../../src/wizard/prompts.js";
 import { createWizardPrompter as createBaseWizardPrompter } from "./wizard-prompter.js";
 
@@ -30,51 +32,52 @@ export function createWizardPrompter(
 }
 
 /** Create isolated auth state and agent directories for auth tests. */
-export async function setupAuthTestEnv(
-  prefix = "openclaw-auth-",
-  options?: { agentSubdir?: string },
-): Promise<{
+type AuthTestEnv = {
   stateDir: string;
   agentDir: string;
-}> {
-  const stateDir = await makeTempWorkspace(prefix);
-  const agentDir = path.join(stateDir, options?.agentSubdir ?? "agent");
-  process.env.OPENCLAW_STATE_DIR = stateDir;
-  process.env.OPENCLAW_AGENT_DIR = agentDir;
-  await fs.mkdir(agentDir, { recursive: true });
-  return { stateDir, agentDir };
-}
-
-type AuthTestLifecycle = {
-  setStateDir: (stateDir: string) => void;
   cleanup: () => Promise<void>;
 };
 
-/** Capture env and track one state dir for cleanup. */
+export async function setupAuthTestEnv(
+  prefix = "openclaw-auth-",
+  options?: { agentSubdir?: string },
+): Promise<AuthTestEnv> {
+  clearRuntimeAuthProfileStoreSnapshots();
+  const state = await createOpenClawTestState({ prefix, layout: "state-only" });
+  const agentDir = path.join(state.stateDir, options?.agentSubdir ?? "agent");
+  process.env.OPENCLAW_AGENT_DIR = agentDir;
+  await fs.mkdir(agentDir, { recursive: true });
+  return { stateDir: state.stateDir, agentDir, cleanup: state.cleanup };
+}
+
+type AuthTestLifecycle = {
+  track: (env: AuthTestEnv) => void;
+  cleanup: () => Promise<void>;
+};
+
+/** Capture env and track isolated OpenClaw state fixtures for cleanup. */
 export function createAuthTestLifecycle(envKeys: string[]): AuthTestLifecycle {
   const envSnapshot = captureEnv(envKeys);
-  let stateDir: string | null = null;
+  const cleanups: Array<() => Promise<void>> = [];
   return {
-    setStateDir(nextStateDir: string) {
-      stateDir = nextStateDir;
+    track(env) {
+      cleanups.push(env.cleanup);
     },
     async cleanup() {
-      if (stateDir) {
-        await fs.rm(stateDir, { recursive: true, force: true });
-        stateDir = null;
+      clearRuntimeAuthProfileStoreSnapshots();
+      for (const cleanup of cleanups.splice(0).reverse()) {
+        await cleanup();
       }
       envSnapshot.restore();
     },
   };
 }
 
-/** Resolve the auth profile JSON path for an agent directory. */
-function authProfilePathForAgent(agentDir: string): string {
-  return path.join(agentDir, "auth-profiles.json");
-}
-
-/** Read and parse auth profiles for an agent directory. */
+/** Read auth profiles from the real SQLite-backed persistence owner. */
 export async function readAuthProfilesForAgent<T>(agentDir: string): Promise<T> {
-  const raw = await fs.readFile(authProfilePathForAgent(agentDir), "utf8");
-  return JSON.parse(raw) as T;
+  const store = loadPersistedAuthProfileStore(agentDir);
+  if (!store) {
+    throw new Error(`Expected SQLite auth profile store for ${agentDir}`);
+  }
+  return store as T;
 }
